Put the combo command into its own module. Updated README and windows_setup script to include the Imgur API Client ID and how to obtain it.
