+++ README.md
@@ -43,11 +43,15 @@ In the repo folder create a file called ``` .env ``` with the contents
 DISCORD_TOKEN=your-discord-bot-token-goes-here
 CLIENT_ID=your-discord-client-id-goes-here
 YOUTUBE_API_KEY=your-youtube-api-key-goes-here
+IMGUR_CLIENT_ID=your-imgur-client-id-goes-here
 ```
 
 To get a Discord bot token and Discord client ID see [How to Make a Discord Bot in the Developer Portal](https://realpython.com/how-to-make-a-discord-bot-python/#how-to-make-a-discord-bot-in-the-developer-portal).
 
 To get a YouTube API key follow the steps outlined [here](https://developers.google.com/youtube/v3/getting-started).
+
+To get an Imgur Client ID, you simply need an Imgur account. Register your bot (application) [here](https://api.imgur.com/oauth2/addclient) and fill out the form.
+**Application Name** and **Email** are required fields. Under **Authorzation Type**, select *OAuth 2 authorization **without** a callback URL*. Imgur will then supply you with a fresh Client ID and Client Secret, but you only really need the Client ID. If you forget it or misplace it, your can access it again in you [Imgur Account Settings](https://imgur.com/account/settings/apps).
 
 After following the steps outlined in [How to Make a Discord Bot in the Developer Portal](https://realpython.com/how-to-make-a-discord-bot-python/#how-to-make-a-discord-bot-in-the-developer-portal), you need to invite the bot to the Discord guilds you would like it to operate in. To do this you will need to navigate to your application in the [Discord Developer Portal](https://discord.com/developers/applications) and then to the ``` OAuth2``` tab.
