+++ windows_setup.bat
@@ -5,6 +5,7 @@ ECHO creating .env file...
     ECHO DISCORD_TOKEN=your-discord-bot-token
     ECHO CLIENT_ID=your-discord-client-id
     ECHO YOUTUBE_API_KEY=your-youtube-api-key
+    ECHO IMGUR_CLIENT_ID=your-imgur-client-id
     ECHO LOGGING=0
 ) > .env
 ECHO npm install
